+++ demo/src/main/java/com/orange/ods/demo/ui/components/ComponentsScreen.kt
@@ -19,7 +19,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -32,7 +31,6 @@ import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.LocalMainTopAppBarManager
 import com.orange.ods.utilities.extension.orElse
 
-@ExperimentalMaterialApi
 @Composable
 fun ComponentsScreen(onComponentClick: (Long) -> Unit) {
     LocalMainTopAppBarManager.current.updateTopAppBarTitle(R.string.navigation_item_components)
@@ -59,7 +57,6 @@ fun ComponentsScreen(onComponentClick: (Long) -> Unit) {
     }
 }
 
-@ExperimentalMaterialApi
 @Composable
 private fun RowScope.ComponentCard(component: Component, onComponentClick: (Long) -> Unit) {
     OdsSmallCard(