+++ demo/src/main/java/com/orange/ods/demo/ui/components/checkboxes/ComponentCheckboxes.kt
@@ -27,7 +27,7 @@ import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomS
 import com.orange.ods.demo.ui.utilities.composable.CheckboxListItem
 import com.orange.ods.demo.ui.utilities.composable.SwitchListItem
 
-@ExperimentalMaterialApi
+@OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun ComponentCheckboxes() {
     val enabled = rememberSaveable { mutableStateOf(true) }