+++ lib/src/main/java/com/orange/ods/compose/component/card/OdsCard.kt
@@ -16,7 +16,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.orange.ods.compose.theme.OdsTheme
 
-@ExperimentalMaterialApi
+@OptIn(ExperimentalMaterialApi::class)
 @Composable
 internal fun OdsCard(modifier: Modifier, onClick: (() -> Unit)?, content: @Composable () -> Unit) {
     val backgroundColor = OdsTheme.colors.surface